pyproject.toml
[tool.poetry]
name = "photo-bank-backend"
version = "0.1.0"
description = ""
authors = ["Zavodchane"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
redis = "^5.0.3"
gunicorn = "^21.2.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
sqlalchemy = {extras = ["asyncio"], version = "2.0.28"}
asyncpg = "^0.29.0"
exif = "^1.6.0"
python-dateutil = "^2.9.0.post0"
pillow = "^10.3.0"
colorthief = "^0.2.1"
aioboto3 = "^12.3.0"
python-multipart = "^0.0.9"
celery = "^5.3.6"
flower = "^2.0.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
uvicorn = "^0.29.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode="auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]